requirements.txt
streamlit
google.generativeai
python-dotenv
json
os
